+++ src/deps/mod.rs
@@ -22,10 +22,8 @@ impl InMemoryStorage {
         system_contracts_options: &crate::system_contracts::Options,
         use_evm_emulator: bool,
     ) -> Self {
-        let contracts = crate::system_contracts::get_deployed_contracts(
-            system_contracts_options,
-            use_evm_emulator,
-        );
+        let contracts =
+            system_contracts::get_deployed_contracts(system_contracts_options, use_evm_emulator);
 
         let system_context_init_log = get_system_context_init_logs(chain_id);
